STarted working on MO1 - Mostly help from copilot. Pushing these here so we at least have some ground to work with as I think this basically covers what we're doing. Feel free to look at it and add/edit to fit the specs better.
I also added the default config.txt which u have to initialize with to the program once u run it via the 'initialize' command. The config is basically the specs for the scheduler. Make sure you have the config.txt in your directory when you pull this too so it can find it and initialize it easily.